+++ server/src/routes/user.ts
@@ -1,17 +1,7 @@
 import express, { Request, Response } from 'express';
-const userRouter = express.Router();
+import UserController from '../controllers/UserController';
+const router = express.Router();
 
-userRouter.get('/', (req: Request, res: Response) => {
-  const { githubId } = req.session;
-  if (githubId) {
-    res.json({
-      githubId,
-    });
-  } else {
-    res.json({
-      githubId: 'not-logged-in',
-    });
-  }
-});
+router.get('/', UserController.login);
 
-export default userRouter;
+export default router;
